Time: 1 ms (23.34%) | Memory: 8.6 MB (13.35%) - LeetSync

+++ 69-sqrtx/sqrtx.cpp
@@ -1,7 +1,8 @@
 class Solution {
 public:
     int mySqrt(int x) {
-        if (x == 0 || x == 1) return x;
+        if (x == 0 || x == 1) 
+            return x;
 
         int l = 1, r = x;
         int ans = 0;
@@ -9,10 +10,10 @@ public:
         while (l <= r) {
             int mid = l + (r - l) / 2;
 
-            if ((long long)mid * mid == x)
+            if ((long)mid * mid == x)
                 return mid;
 
-            else if ((long long)mid * mid < x) {
+            else if ((long)mid * mid < x) {
                 ans = mid;
                 l = mid + 1;
             }
